pyproject.toml
[tool.poetry]
name = "dj-celery-rabbitmq"
version = "0.1.0"
description = ""
authors = ["AlekysCoelho <alekyscoelho@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
celery = "^5.4.0"
django-ninja = "^1.3.0"
yfinance = "^0.2.43"
django-celery-results = "^2.5.1"
django-celery-beat = "^2.7.0"


[tool.poetry.group.dev.dependencies]
python-decouple = "^3.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
